Measure time correctly for download speed. (#64)
* Measure time correctly for download speed.

Instead of adding up all the time spent downloading, we need to find the
union of all the intervals spent downloading otherwise the total
download time could be longer than the build.

Fixes #63

+++ doctor-plugin/src/main/java/com/osacky/doctor/internal/SlowNetworkPrinter.kt
@@ -2,7 +2,7 @@ package com.osacky.doctor.internal
 
 class SlowNetworkPrinter(private val type: String) {
 
-    fun obtainMessage(totalBytes: Int, totalTime: Int, totalSpeed: Float): String {
+    fun obtainMessage(totalBytes: Int, totalTime: Long, totalSpeed: Float): String {
         val megabytesDownloaded = twoDigits.format(totalBytes * 1.0f / ONE_MEGABYTE)
         val secondsDownloading = twoDigits.format(totalTime * 1.0f / 1000)
         val totalSpeedFormatted = twoDigits.format(totalSpeed)
